* Port [3899] to 1.2.16 from trunk  * Cleanup 64-bit warnings/errors
svn: 3900

+++ src/crypto/cleanse.c
@@ -63,7 +63,7 @@ void OPENSSL_cleanse(void *ptr, size_t len)
         while(loop--)
                 {
                 *(p++) = cleanse_ctr;
-                cleanse_ctr += (17 + (unsigned char)((int)p & 0xF));
+                cleanse_ctr += (17 + (unsigned char)((long)p & 0xF));
                 }
         if(memchr(ptr, cleanse_ctr, len))
                 cleanse_ctr += 63;
